load application state at start
