+++ server/actions/client.js
@@ -1,0 +1,8 @@
+export function loadState() {
+  return (dispatch, getState) => {
+    dispatch({
+      type: "LOAD_STATE",
+      state: getState()
+    });
+  };
+}
